feat: добавлены эндпоинты для админа и для stats

+++ ewm/src/main/java/ru/practicum/explorewithme/event/dto/AdminUpdateEventRequest.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class AdminUpdateEventRequest {
     private String annotation;
-    private int category;
+    private long category;
     private String description;
     private LocalDateTime eventDate;
     private boolean paid;
